requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
Faker==38.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
